% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textLikelihood.R
\name{textLikelihood}
\alias{textLikelihood}
\title{Likelihood of a string}
\usage{
textLikelihood(text, log_scale = TRUE)
}
\arguments{
\item{text}{A character vector of length 1.}

\item{log_scale}{If \code{TRUE} (default) log-likelihood is returned.}
}
\value{
A numeric(1) of the total (log-)likelihood of the input text.
}
\description{
Likelihood of a string
}
